3161: Refactored ChatMenu
