+++ web/src/components/ChatContainer.tsx
@@ -13,7 +13,7 @@ import buildConfig from '../constants/buildConfig'
 import useDimensions from '../hooks/useDimensions'
 import useLocalStorage from '../hooks/useLocalStorage'
 import useLockedBody from '../hooks/useLockedBody'
-import { chatIdKey } from '../utils/chatId'
+import { chatIdKey } from '../utils/chat'
 import ChatController from './ChatController'
 import ChatMenu from './ChatMenu'
 import HeaderLanguageSelectorItem from './HeaderLanguageSelectorItem'
