+++ web/src/utils/chat.ts
@@ -1,0 +1,5 @@
+export const LOCAL_STORAGE_ITEM_CHAT_ID = 'Chat-Device-Id'
+
+export const chatIdKey = (cityCode: string): string => `${LOCAL_STORAGE_ITEM_CHAT_ID}-${cityCode}`
+
+export const generateChatId = (): string => globalThis.crypto.randomUUID()
